+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Next Release
 
+## 0.0.10
+
+- [router, sdk] Refactor error handling.
+- [sdk] [#53](https://github.com/connext/nxtp/issues/53) Add `ReceiverPrepareSigned` event.
+- [sdk] [#54](https://github.com/connext/nxtp/issues/54) Fix `finishTransfer` resolution.
+
 ## 0.0.9
 
 - [sdk] Add methods to attach and detach listener.